requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
future==0.18.3
idna==3.6
line-bot-sdk==1.8.0
lxml==4.9.4
pytz==2023.3.post1
requests==2.31.0
schedule==1.2.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
